refactor: synchronize project resource loading
